requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.2
brotlipy==0.7.0
bs4==0.0.1
certifi==2023.5.7
cffi
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.1.1
cpm-kernels==1.0.11
cryptography
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.14
dnspython==2.4.2
exceptiongroup==1.1.3
faiss-cpu==1.7.0
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.4
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.9.1
gradio==3.44.3
gradio_client==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.1
hyperlink==21.0.0
idna 
importlib-metadata==6.6.0
importlib-resources==6.0.1
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.273
langsmith==0.0.26
latex2mathml==3.76.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.3
mdtex2html==1.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
openai==0.27.6
OpenCC==1.1.6
orjson==3.9.7
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsel==1.8.1
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pydantic==1.10.9
pydantic_core==2.6.3
PyDispatcher==2.0.7
pydub==0.25.1
PyJWT==2.8.0
pymongo==4.5.0
pyOpenSSL 
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.8.8
requests==2.30.0
requests-file==1.5.1
rpds-py==0.10.3
scikit-learn==1.3.0
scipy==1.10.1
Scrapy==2.11.0
scrapy-splash==0.9.0
scrapy-user-agents==0.1.1
selenium==4.12.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tldextract==5.0.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.27.1
trio==0.22.2
trio-websocket==0.10.4
Twisted==22.10.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.6
user-agents==2.2.0
uvicorn==0.23.2
w3lib==2.1.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
zope.interface==6.1
